+++ src/main/java/roomescape/reservation/presentation/dto/ReservationInfo.java
@@ -1,8 +1,8 @@
-package roomescape.controller.dto;
+package roomescape.reservation.presentation.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public record ReservationInfo(
         Long id,
