+++ src/main/java/roomescape/reservation/presentation/dto/ReservationRequest.java
@@ -1,9 +1,9 @@
-package roomescape.controller.dto;
+package roomescape.reservation.presentation.dto;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import roomescape.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public record ReservationRequest(
         String name,
