+++ src/test/java/roomescape/reservation/presentation/MissionStepTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.reservation.presentation;
 
 import static org.hamcrest.Matchers.is;
